+++ HISTORY.md
@@ -1,5 +1,12 @@
 # DynamicPPL Changelog
 
+## 0.38.8
+
+Added a new exported struct, `DynamicPPL.ParamsWithStats`.
+This can broadly be used to represent the output of a model: it consists of an `OrderedDict` of `VarName` parameters and their values, along with a `stats` NamedTuple which can hold arbitrary data, such as (but not limited to) log-probabilities.
+
+Implemented the functions `AbstractMCMC.to_samples` and `AbstractMCMC.from_samples`, which convert between an `MCMCChains.Chains` object and a matrix of `DynamicPPL.ParamsWithStats` objects.
+
 ## 0.38.7
 
 Made a small tweak to DynamicPPL's compiler output to avoid potential undefined variables when resuming model functions midway through (e.g. with Libtask in Turing's SMC/PG samplers).